+++ src/eslint/rules/base.js
@@ -21,6 +21,7 @@ module.exports = {
   plugins: ["react", "react-hooks", "jsx-a11y"],
 
   extends: [
+    "airbnb",
     "prettier",
     "eslint:recommended",
     "plugin:react/recommended",