+++ src/iqws.js
@@ -1,17 +1,20 @@
 #!/usr/bin/env node
 
-"use strict";
+'use strict';
+
+/* eslint-disable no-console */
 
 // Makes the script crash on unhandled rejections instead of silently
 // ignoring them. In the future, promise rejections that are not handled will
 // terminate the Node.js process with a non-zero exit code.
-process.on("unhandledRejection", err => {
+process.on('unhandledRejection', err => {
   throw err;
 });
 
-const crossSpawn = require("cross-spawn");
+const crossSpawn = require('cross-spawn');
+
 const args = process.argv.slice(2);
-const validScripts = ["lint", "linc", "prettier", "help"];
+const validScripts = ['lint', 'linc', 'prettier', 'help'];
 
 // check for valid commands
 let scriptIndex = args.findIndex(x => validScripts.includes(x));
@@ -22,26 +25,30 @@ const nodeArgs = scriptIndex > 0 ? args.slice(0, scriptIndex) : [];
 const extraArgs = args.slice(scriptIndex + 1);
 
 // alias for help script
-if (["-h", "--help"].includes(script)) {
+if (['-h', '--help'].includes(script)) {
   scriptIndex = 0;
-  script = "help";
+  script = 'help';
 } else if (validScripts.includes(script)) {
   // check if it using a valid script
-  const result = crossSpawn.sync("node", nodeArgs.concat(require.resolve("./bin/" + script)).concat(extraArgs), {
-    stdio: "inherit",
-  });
+  const result = crossSpawn.sync(
+    'node',
+    `${nodeArgs} ${require.resolve(`./bin/${script}`)} ${extraArgs}`,
+    {
+      stdio: 'inherit',
+    }
+  );
   if (result.signal) {
-    if (result.signal === "SIGKILL") {
+    if (result.signal === 'SIGKILL') {
       console.log(
-        "The build failed because the process exited too early. " +
-          "This probably means the system ran out of memory or someone called " +
-          "`kill -9` on the process."
+        'The build failed because the process exited too early. ' +
+          'This probably means the system ran out of memory or someone called ' +
+          '`kill -9` on the process.'
       );
-    } else if (result.signal === "SIGTERM") {
+    } else if (result.signal === 'SIGTERM') {
       console.log(
-        "The build failed because the process exited too early. " +
-          "Someone might have called `kill` or `killall`, or the system could " +
-          "be shutting down."
+        'The build failed because the process exited too early. ' +
+          'Someone might have called `kill` or `killall`, or the system could ' +
+          'be shutting down.'
       );
     }
     process.exit(1);
@@ -49,5 +56,5 @@ if (["-h", "--help"].includes(script)) {
   process.exit(result.status);
 } else {
   console.log(`Unknown script ${script}`);
-  console.log("Perhaps you need to update iq-web-scripts?");
+  console.log('Perhaps you need to update iq-web-scripts?');
 }