+++ src/prettier/index.js
@@ -1,5 +1,7 @@
 "use strict";
 
+/* eslint-disable no-console */
+
 // Based on similar script in Jest
 // https://github.com/facebook/jest/blob/a7acc5ae519613647ff2c253dd21933d6f94b47f/scripts/prettier.js
 
@@ -8,6 +10,7 @@ const glob = require("glob");
 const prettier = require("prettier");
 const fs = require("fs");
 const listChangedFiles = require("../shared/listChangedFiles");
+
 const prettierConfigPath = require.resolve("./.prettierrc");
 
 const mode = process.argv[2] || "check";
@@ -18,13 +21,7 @@ const changedFiles = onlyChanged ? listChangedFiles() : null;
 let didWarn = false;
 let didError = false;
 
-const filesToIgnore = [
-  "**/node_modules/**",
-  "**/dist/**",
-  "**/package.json",
-  "**/tsconfig.json",
-  "**/webpack.*",
-];
+const filesToIgnore = ["**/node_modules/**", "**/dist/**", "**/package.json", "**/tsconfig.json", "**/webpack.*"];
 
 const runPrettier = fileList => {
   fileList.forEach(file => {
@@ -38,25 +35,21 @@ const runPrettier = fileList => {
         if (output !== input) {
           fs.writeFileSync(file, output, "utf8");
         }
-      } else {
-        if (!prettier.check(input, options)) {
-          if (!didWarn) {
-            console.log(
-              "\n" +
-                chalk.red("  This project uses prettier to format all JavaScript code.\n") +
-                chalk.dim("    Please run ") +
-                chalk.reset("prettier --write") +
-                chalk.dim(" and add changes to files listed below to your commit:") +
-                "\n\n"
-            );
-            didWarn = true;
-          }
-          console.log(file);
+      } else if (!prettier.check(input, options)) {
+        if (!didWarn) {
+          console.log(`\n\r
+            \r${chalk.red("  This project uses prettier to format all JavaScript code.")}
+            \r${chalk.dim("  Please run")} ${chalk.reset("prettier --write")}
+            \r${chalk.dim("  and add changes to files listed below to your commit:")}
+              
+          `);
+          didWarn = true;
         }
+        console.log(file);
       }
     } catch (error) {
       didError = true;
-      console.log("\n\n" + error.message);
+      console.log(`\n\n ${error.message}`);
       console.log(file);
     }
   });