+++ src/shared/evalToString.js
@@ -7,11 +7,11 @@ function evalToString(ast /* : Object */) /* : string */ {
       return ast.value;
     case "BinaryExpression": // `+`
       if (ast.operator !== "+") {
-        throw new Error("Unsupported binary operator " + ast.operator);
+        throw new Error(`Unsupported binary operator ${ast.operator}`);
       }
       return evalToString(ast.left) + evalToString(ast.right);
     default:
-      throw new Error("Unsupported type " + ast.type);
+      throw new Error(`Unsupported type ${ast.type}`);
   }
 }