+++ changelog.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 06/16/2021
+
+### Updated
+
+- Dependencies
+  - All ESlint plugins
+  - Prettier
+
+### Added
+
+- Direct extension of Airbnb style guide
+- Additional rules for react and typescript from `eslint-config-react-app`
+
 ## [1.0.2] - 10/12/2020
 
 ### Fixed